Set SameSite=None for licensify frontend
Some payment providers send responses using an HTML form which POSTs
data  from the user's browser to Licensing. For this to work, the POST
request  needs to include the session cookie. This means the SameSite
attribute on the  session cookie has to be set to None (rather than the
default which is "Lax"  as of 2021).

https://web.dev/samesite-cookies-explained/

Note that this commit only touches the config file for licensify
frontend - licensify-admin and licensify-feed have separate .conf.erb
files in the same directory as this.

This depends on https://github.com/alphagov/licensify/pull/514 which
upgraddes Play to 2.6.25