Add a global Icinga check for checking TCP connection
This is being added as a means to check whether a TCP port is open for
services. The use case I'm adding this for is to check whether a Docker
server is running (as per the precedent [1]) and creating a generalised
means to check for it so we don't have lots of checks for each type of
Docker server.

There's probably better ways to check this by querying Docker directly
but this should be sufficient for our needs.

[1]: https://github.com/alphagov/govuk-puppet/commit/72e81fb488dd32bc52797750c7866c8302c84b61