Remove confusing contacts hieradata from frontend
It looks like this configuration was added in
6486c50e26a71299c32c458d6fdde5c289e1a20a, but it's a little confusing
even back then because the `contacts` application doesn't run on the
`frontend` node, so surely these settings would never have taken effect?

I'm removing them now because I'd like to remove the mysql-replica RDS
instances, but this configuration makes it look like they're still used.

I'm pretty confident they're not used however, because:

1) The CloudWatch metrics graph of ConnectionCount for mysql-replica is
consistently 0
2) `gds govuk connect -e integration ssh frontend ls /var/apps` confirms
that `contacts` is not present on the frontend node class
3) `gds govuk connect -e integration ssh backend cat
/etc/govuk/contacts/env.d/DATABASE_URL` confirms that the backend node
class is using mysql-primary (not mysql-replica)