Put conditionals into alphabetical order
We'll soon be creating a number of different {app}_db_admin
machines, all of which will require their own conditional in this
`dump_postgresql` function. We should start as we mean to go on,
and put these in an alphabetical order to ease maintenance.

Note that in future we hope to refactor away the need to have
these conditionals.