Configure govuk_env_sync to accept a database hostname
This replaces the conditionals govuk_env_sync with a database hostname
argument (-H or --database_hostname).

This has been done to remove the conditionals needed to map database to
a particular hostname - which would be to grow in size with the current
work to create new RDS instances.

A motivation for this work is to allow us to configure new database RDS
instances to pull from existing databases so they can be populated
automatically.

In the template I found following the convention of not prefixing a
variable with an @ didn't work (it treated each execution as true) so I
switched it to use an @. I also made the output a bit neater by using
the -%> to cull some unnecessary new line characters.