DRY up the app-specific DB admin box configuration
We're about to create many app-specific DB admin boxes, all of
which will have the same configuration on the node itself, so it
makes sense to encapsulate this in one place.

I did consider leaving the existing db_admin node untouched, as
we will eventually remove the Postgres code from it when every
Postgres-powered app is using its own RDS instance. However, it
uses the same code we've just encapsulated (albeit in a slightly
different order), so there's little harm in making this optimisation
now.