Add Icinga alerting to Ask Export
This add Icinga notifying into the process of the job completing or
failing, without this we won't actually receive any alerts if the
service fails or stops running - the check will just remain as a forever
pending state.

I've changed the passive check from being conditional to only having the
freshness check being conditional - this is so this can still alert in
integration and having the alerting tested there.