Move .my.cnf.erb to root of templates directory
We're going to reuse the MySQL .my.cnf contents for all of the
DB admin machines we're about to create for the MySQL RDS instances,
as the configuration is the same for each. Having these new
app-specific DB admin machines make a reference to the shared
s_db_admin node could be confusing, so we've renamed the template
to clarify that it is not only used by s_db_admin.