Backup and restore CKAN PYCSW database
Today we had the surprise of learning that CKAN actually has a second
PostgreSQL database for PYCSW (which I assume is: https://pycsw.org/),
it turns out we don't actually backup and restore from that database.

This therefore sets that configuration up, I've tried to roughly follow
existing conventions on timings - but they seem mostly arbitrary.