Don't sync account-api data from production to staging / integration
The account-api database contains real user data (eg email addresses),
and we haven't implemented a sanitisation step (like we have for
email-alert-api).  Furthermore, this data is of limited use in other
environments as the subject identifiers won't match the non-production
Digital Identity service.

So we don't sync this data to staging and integration.