Always create env.d dir for govuk_env_sync module
This directory is only created if there's a task.  But not all of the
db-admin machines have any tasks right now: a lot of them are
commented out.

This is a problem because the document_db_credentials value is set in
govuk-secrets, so the govuk_env_sync::documentdb_auth resource gets
pulled in, and that expects this directory to exist.

Since not creating the directory seems of little value, always create
it, even if nothing needs it.