Change postgres hostnames to DATABASE-NAME-postgres
The hostnames are generated based on the database name and the AWS RDS
engine, and AWS calls the postgres engine "postgres", not
"postgresql".

Changing the hostnames to be DATABASE-NAME-postgresql would require
either specifying the hostname of every database in govuk-aws-data, or
special-casing the postgres hostnames.  Neither are particularly good,
and I think just using the names we currently have is better.