Stop auto-creating MySQL db/users on the new DB admin machines
See the new 'create-mysql-db-and-users.md' file for an explanation
as to why we've stopped attempting to create the databases and
users automatically.

This will now be a manual process. A link to where the document
will live in the GOV.UK Developer Docs has been added to each
relevant db_admin definition, for ease of finding.

Whilst this renders each app's `db.pp` file essentially useless
at this point, I was not able to delete the files because they are
required by `s_db_admin.pp` and `s_backup.pp`. It is safer to
simply keep the `db.pp` files until we've fully migrated to
the RFC-143 approach.