Allow govuk-backup to run `sudo docker`
The docker socket is only writeable by root/docker, and the govuk-backup
user is not in the docker group. This means the user can't do `docker
run` or similar commands, because the docker client needs to be able to
write to the docker socket for that to work.

Running docker as `sudo` works around this, as it's then able to write
to the socket. However, the govuk-backup user wasn't allowed to `sudo
docker`. This commit allows that.

This potentially enables govuk-backup to escalate its privileges to
root, since it can do things like:

```
docker run -it -v /:/host ubuntu
```

(i.e. run a ubuntu container [as root], and mount the entire filesystem as
a volume)

This doesn't feel like a sensible thing to be worried about because the
govuk-backup user can already probably escalate its privileges with the
other commands in the sudoers file (e.g. psql can be used to overwrite
local files, and it can run psql with sudo). Also all of this is going
away soon with the move to k8s.