Add Jenkins job for Search API's learn to rank
This adds a Jenkins job to train and deploy Search API's machine
learning model for ranking results (learn to rank). This replicates the
previous pipeline setup in concourse as GDS concourse is being shut
down. The job fetches training data from Search API's elasticsearch and
GCP Bigquery to create a training dataset, which is stored in AWS S3.
Then the job triggers AWS SageMaker to train a model based on that
dataset and deploy the new model.