Comment out pull_ckan_daily env sync tasks
These are currently not working due to the postgres 13 upgrade -
the old pg_restore client doesn't like restoring dumps created by the
new pg_dump thing.

It's a bit involved fixing the script, and it's christmas time. So we're
just going to stop the restore bit of the sync process for now and come
back to it in the new year.