Temporarily disable Publishing API env sync tasks in staging
Over the weekend we're going to be running an AWS Database Migration
Service sync between the postgresql-primary RDS instance to the
publishing-api-postgres RDS instance. The env sync process would break
this job as it drops and rebuilds databases. Therefore it is to be
disabled.

I plan to reinstate these jobs on Monday following the migration.