Remove Postgres config from db_admin machine
Postgres is no longer required on the general db_admin machine and so we
need to mark this as absent, before removing the code entirely. We do
use the postgresql_db_admin class to configure the app specific db_admin
machines so we need to mark this as present by default.