Make Smokey job names more descriptive
Following on from the success of #11455, we can make a similar
change to our Smokey builds, so that it is easy to see whether
a build was for the entire test suite or for a particular app.

When TARGET_APPLICATION is set, it displays in the job name,
otherwise we have an empty string. I couldn't figure out how to
set it to a default, e.g. 'All'.