Update branch for the LTR jenkins job
This updates the branch reference to 'main' instead of the development
branch. This job previous referenced the development branch, as to all
the concourse pipeline to still function concurrently whilst the
migration took place. However, now concourse pipeline is no longer
needed and changes have been merged in to main.