Remove PostgreSQL configuration from DB Admin machines
This is an alternative to: https://github.com/alphagov/govuk-puppet/pull/11461
which used Puppet to try remove PostgreSQL and it's config.

We've since learnt that this isn't possible, if `postgresql::server`
gets called with `package_ensure => absent` it will still try to create
directories [1] and thus puppet will error.

Therefore I am taking a simpler option here and just removing the config
with an intention to recycle the boxes to remove the stuff that isn't
needed anymore (which I'll plan to do after the MySQL configuration is
removed).

[1]: https://github.com/puppetlabs/puppetlabs-postgresql/blob/6a07decd104dc1e004135bae6403e516cf13bbe2/manifests/server/initdb.pp#L24-L29