Stop dumping staging databases to S3
The template script which was used to set up DB Admin machines included a "push" env_sync task to dump staging databases into a staging S3 bucket.

It's since become apparent that, in most cases, these staging dumps aren't used and are therefore not needed. This is because all integration environments (with the exception of Whitehall) pull data directly from the production S3 bucket.