Add dashboard for Publishing Service
This adds a dashboard for tracking metrics relevant to the publishing
service team. In particular this shows 'fact check emails alert', which allows
us to delete the icinga alert for this problem. It requires the service
team to monitor this dashboard/problem, in keeping with the devops/service model
the organisation is moving towards.

In another commit I will remove the fact check emails alert that 2nd line is paged with.

Temporary view of dashboard available at:
https://grafana.blue.production.govuk.digital/dashboard/db/temporary-publishing-service?orgId=1&from=now-1h&to=now