Add automated-rake-task job
This duplicates the run-rake-task Jenkins job which has been
run 225,038 times in one environment.

We'd like to see what jobs are run manually vs automated so
we have a better idea of the necessity for a manual job runner
and the features we'll need to implement to replace the manual
rake tasks.

A following commit will change all the automated jobs to
trigger the automated-rake-task job so we don't need to update
our docs / manual workflow to use a different job.