Only do daily backups of MongoDB in integration
We only run this script in integration and it only seems to have the
purpose of populating the files that are used for a govuk-docker
replication task [1] (which itself is a bit weird, it's unclear why it
doesn't use env sync backups like everything else... but yaks).

Because of this it seems rather superfluous to be generating backups
every 30 mins.

This changes the script to only run once a day, it is set to run at 8am
to be out the way of the database syncs.

I simplified the test a bit as it seemed unnecessary for the test to be
coupled with the time this task runs.

[1]: https://github.com/alphagov/govuk-docker/blob/6004cdc44ff60fc59e9b7aa156737621ff780a6d/bin/replicate-mongodb.sh#L52-L54