Remove cache invalidation emergency banner downstream jobs
This commit removes the cache invalidation logic from the
add/remove emergency banner jobs, in order to implement RFC 144 [1].

We no longer need to clear the caches as the TTL on GOV.UK pages has
been reduced to 5 minutes. We can therefore wait 5 minutes for the banner
to appear on all GOV.UK pages.

I have kept the jobs to purge the varnish / Fastly caches in Jenkins,
as we may need them in future.

[1]: https://github.com/alphagov/govuk-rfcs/blob/main/rfc-144-consistent-cache-expiry.md