Generate Jenkins' public ssh key
Previously, the govuk_jenkins ssh_key class required you to pass in the
private key and the public key separately.

This meant configuring the public keys in hieradata_aws, and the private
keys in govuk-secrets, and making sure you got the right public keys for
the right private keys.

Predictably, we did not get this right:

* https://github.com/alphagov/govuk-puppet/pull/11543
* https://github.com/alphagov/govuk-aws-data/pull/980

There's no need to ask for both the private key and the public key.
Since the private key _contains_ the public key we can just generate it.

The only reason I can think of for specifying them separately would be
if we wanted to allow the mistake of having mismatched keys.

Using Puppet's Exec resource, we can run `ssh-keygen` to create the
public key from the private key. The `creates` property ensures this
runs if the key is not already there.

Using the `subscribe` property means that the exec resource should be
notified if the private key resource changes, and run the command again.