Ship /var/log/auth.log to cloudwatch logs
This creates a simple configuration file which will ship the contents of
/var/log/auth.log (which includes details of who's ssh'd into a machine,
as well as things like which sudo commands have been run) to cloudwatch
logs.

The service configuration ensures that the amazon-cloudwatch-agent will
be restarted if the configuration file changes.

The JSON was created by running amazon-cloudwatch-agent-config-wizard on
the jumpbox, and it's been tested with a snowflaked jumpbox on
integration.

The intention is to further ship these logs from cloudwatch into splunk,
so that the cyber security team have some visibility of these events.