Install ChromeDriver alongside Google Chrome
This adds a module `govuk_chromedriver` which can be included to install Google Chrome alongside a compatible version of ChromeDriver, ready for use by Selenium browser tests.

This is being introduced so that there is always a usable version of ChromeDriver available on CI machines which need to run tests. In the past, CI machines have had a system-wide version of Google Chrome available for use, but no accompanying ChromeDriver. Instead each application has had to use the `webdrivers` gem to install its own version of ChromeDriver locally.

This change represents a deliberate shift in responsibility. Where previously applications were responsible for sourcing their own ChromeDriver (and hence depended on the `webdrivers` gem), we've now decided that it be the responsibility of the underlying operating environment to provide it.

In real terms, this means that the GOV.UK development environment – [govuk-docker] – and the CI environment need to have ChromeDriver installed.

[govuk-docker]: https://github.com/alphagov/govuk-docker/pull/577