FIx govuk_crawler script
We were trying to use globbing on mkdir -p to get around having to work out
what the base mirror directory was called. It's environment dependent and
is the website hostname, so www.gov.uk and www.staging.publishing.service.gov.uk
on production and staging respectively.

We were trying to get around this by using www.* and hoping it would expand
to the right thing, so that we could create an error subdirectory in whatever
existed.

The first problem was that the /mnt/crawler_worker directory (and children)
are owned by the deploy user, so the creation failed. The second is that
the `mkdir -p www.*/error/`command created a directory called (literally)
`www.*`, which is not what we want at all.

So, I've switched this around to manually get the directory name from the
`GOVUK_WEBSITE_ROOT` environment variable and use it to construct the path.

We then need to use the deploy user to create the directory and grab the
error file from static.