Remove Selenium Standalone Server from test boxes
The Selenium Standalone Server, which was installed by the `selenium` module, is no longer required now that we have ChromeDriver installed on test machines.

This commit renames the `selenium` module to `remove_selenium`, and repurposes it to remove existing Selenium Standalone Server binaries.

This achieves the first step of removing Selenium: deleting the binaries. In a later commit, once this change has been applied, we'll be able to remove the configuration entirely as it'll no longer be needed.