Remove Xvfb from govuk_testing_tools module
[Xvfb](https://www.x.org/releases/X11R7.6/doc/man/man1/Xvfb.1.xhtml) is a virtual X server, which was previously needed so capybara-webkit could headlessly render web pages.

Capybara-webkit is no longer used now that we have ChromeDriver installed via the govuk_chromedriver module.

This change removes the Xvfb package and its configuration file. It does this by renaming and repurposing the existing puppet class.

This achieves the first step of removing Xvfb. Once this change has been applied, we'll be able to remove the configuration entirely as it'll no longer be needed.