Remove Bouncer minimum request rate alert
This alert has been around for a long time and just causes noise on
2ndline, with no real value. The minimum requests threshold is arbitrary
and I can't really think of a way in which this is useful, as a low
amount of requests doesn't mean there is a problem.

Also, the frequency we're seeing the alert means that it's likely to be
ignored, even if there was an issue.