Remove content-store from govuk-uptime-collector list
By removing the public load balancer of content-store as per
https://github.com/alphagov/govuk-aws/pull/1572 we also removed
the public hostname for content-store.
This breaks govuk-uptime-collector which depends on this hostname
for the check url of content-store.
The solution is to simply remove content-store of the list of
services monitored by govuk-uptime-collector, it should probably
never have been there in the first place as evidenced by :
https://docs.publishing.service.gov.uk/manual/uptime-metrics.html