Reduce default local log retention to 7 days.
GOV.UK uses Logit for centralised log management and retention, and local logs will disappear with the migration to Kubernetes,
so these logs don't form part of our log retention policy.

This commit brings the default local  log retention down from 31 days to 7, with the primary goal of removing out-of-disk errors
for Content Store. This also makes the default consistent with a couple of existing apps which override the default
(search and cache).

This doesn't affect the local backup retention for other things (e.g.local Mongo).