Manually delete all jobs which aren't in the `$pipeline_jobs` hash
When items are removed from this hash, Puppet doesn't remove the
job directories, so the jobs are never removed from Jenkins. This
script is an attempt to resolve that: after importing all the
relevant jobs, we then remove any directories we don't recognise.

Thanks to Chris B for helping improve the shell script.

Co-authored-by: Chris B <chris.banks@digital.cabinet-office.gov.uk>