Prune rotated files older than rotate value.
This commit introduces a clean-up to delete any left-over rotated files that are older than
the current 'rotate' value.

Under normal circumstances, logrotate will remove logs as they rotate off the end. We set the maximum
number of rotations (in our case days) with the 'rotate' value.

However, an edge-case exists where older files will not be removed if we've reduced the value of rotate
between runs. This is because logrotate only looks at 'rotate + 1' to check for removal, it doesn't
attempt to manage all files in that directory.

This clean-up mitigates that edge-case, ensuring that all files older than the maximum number of days are
removed. There are a couple of caveats:

1. The clean-up uses modification time, so if we touch any older files they won't be removed.
2. We assume that logrotate will continue to be daily. Any other setting will break the clean-up. This is not
   a configurable parameter in our Puppet setup, so this risk feels acceptable