Set 'ENSURE_DEFAULT_BRANCH' ENV variable
Otherwise this leads to a failing build, with an "Unbound local
variable" error on this line:
https://github.com/alphagov/govuk-repo-mirror/blob/91bc353ee40d930476999764a2a172473827c63b/mirror_repos#L74

The value is set to `true` on
https://github.com/alphagov/govuk-repo-mirror/blob/f5e8ef3632b7b070141f82d0b4e66ef77b4ae282/concourse.yml#L53.
However, the current behaviour of the `ensure_aws_default_branch`
script seems broken.

For example, this clearly incorrect default branch for Static:

```
Getting current default branch for static repo
Repo static has default branch: add-template-full_width-and-omit_footer_navigation, updating to 'main'
```

...and this incorrect default of 'master' (rather than 'main')
for Support API (which has had its default branch set to 'main'
for some time):

```
Getting current default branch for support-api repo
Main branch does not exist, will set default branch to master.
```

Re-running the mirroring script didn't change the above outcome.

This 'ensure_aws_default_branch' script is also slow to run,
adding several minutes to the overall build time. It's also not
clear what benefits there are to running it; the only thing that
is clear is that it's "optional":
https://github.com/alphagov/govuk-repo-mirror/blob/083803c2052ef5a239cb1fa5564a704867a3640a/ensure_aws_default_branch#L7

Therefore I've set this to `false` to stop it running, until it
can be fixed and optimised.