Add repository name for Contacts Admin downstream job
Continuous deployment of Contacts Admin is currently failing due to the
following error:

```
Could not resolve to a Repository with the name 'alphagov/contacts'
```

This application's repo name does not match it's deployed name.
Therefore we need to specify the repository name to enable it to be
pulled by the downstream deployment job.