Accept URL parameters in Clear CDN Cache task
This changes the Clear CDN Cache task to accept a collection of URLs as
parameters rather than hard-coding 10 URLs to remove.

This has been added to work towards feature parity with the
cache-clearing service rake tasks [1] which we are considering retiring.

This changes this script to accept a string of space separated URLs

Some other changes from the past script:

- It checks what type of URL is provided to decide what to clear. If
  it's just a path it'll prefix the GOV.UK hostname, but a full URL can
  be used for assets
- Because of the above we can drop passing in the asset hostname as a
  puppet parameter
- We only send the message to Fastly from one machine, I don't see why
  we did it from every node 🤯
- If a curl request fails it'll cause the script to fail

[1]: https://github.com/alphagov/cache-clearing-service/blob/64a4bebe97497102f286bfab6e613e33f61f0183/lib/tasks/cache.rake