Include internal varnish cache on CDN cache clear task
This updates the task to clear individual pages from the CDN cache to
also do the internal varnish cache. I did this because there wasn't an
equivalent task (only one to nuke the entire varnish cache) and it
didn't strike me that there was a downside in doing this.

This has been done for continued parity with the cache clearing service
rake tasks ahead of it's retirement [1].

[1]: https://github.com/alphagov/cache-clearing-service/blob/64a4bebe97497102f286bfab6e613e33f61f0183/lib/tasks/cache.rake#L18-L22