Remove automatic running of Whitehall Link Check report
This is a report sent to publishers (on a monthly basis) that lists all
links that are marked as broken in their published content.

The GOV.UK Publishing Service team have decided to temporarily pause the
automatic sending of this report whilst improvements to its content are
made.

This removes the monthly trigger.