Make logrotate aware of disk space
Delete logs files above a threshold in size in chronological order
while disk space usage is above a given threshold.
We've had several issues with machines running out of disk space
because of oversized log files. This is an attempt to prevent
this from happening, as such it is a blanket change applying to
all rotated logs on all classes.
The thresholds were picked arbitrarily initially, SIZE is set at
300M to avoid getting rid of small files that would not help much
to save space, and MAX_USE is at 85% of use as our nagios
thresholds for disk space are 80% warning, 90% critical.