Update Jenkins Docker plugins.
Update to the latest minor releases of the docker-commons and
docker-workflow plugins.

These require a somewhat newer version of pipeline-model-definition.
I've gone for the oldest version that's compatible with the updated
Docker plugins.

We need this for the fix for
https://issues.jenkins.io/browse/JENKINS-44609 in order to use a build
arg for the Ruby version, which we're going to need in order to
implement our base/builder images so that we can rebuild everything
easily (e.g. when there are security patches in libraries).