Remove redundant fallback for app healthchecks
This was added in [^1] without explanation.

I've checked and all apps have has_readiness_health_check set to
true wherever has_liveness_health_checkset is also set to true,
so this conditional branch is never used in practice.

I think it's clearer to have the alert work with just one type of
healthcheck and require someone adding an app to add both. This
means we can simplify the documentation for the alert.

[^1]: https://github.com/alphagov/govuk-puppet/commit/50c30a600617df809238715cb6c398b60cdea2aa