Fix job status for deploy_terraform_govuk_aws job
It seems that the output of terraform apply has changed with
new version of terraform where there is a plan before apply
so we regex on `Plan ..` to mark a build as unstable, terraform
apply builds are marked as unstable.

The solution is to remove this regex and add a -detailed-exitcode
option to terraform plans so that the exit code is 2 when there
are changes to be applied and Jenkins will interpret exit code 2
as unstable builds.

Ref:
1. [exit code changes](https://github.com/alphagov/govuk-aws/pull/1592)