Ignore "no healthy targets" for Manuals Frontend
Manuals Frontend is in the process of being deprecated.

We currently have a "AWS LB Healthy Hosts" error for this application,
as the load balancer cannot find any healthy targets.

The alert is currently acknowledged (so we don't see it), but as the
alert is shared with all other frontend applications, this means we are
also hiding alerts for other (non-deprecated) applications.

This will hide the alert for just Manuals Frontend, so the alert can be
un-acknowledged.