Push mongo databases to integration backup directory
It seems that in the integration environment we've not configured these
databases to be pushed to the backups folder and thus they can't be used
by developers to replicate data locally.

I've configured these to have a window between the production pull and
the integration dump similar to that of production.