Set account home environment variables
These are used by `govuk_personalisation`[1] to populate links to the
account home page.

We are in the process of moving the account home page from
`www.gov.uk/account/home` (served by Frontend) to an app on the DI side,
which is part of separating out central account features from the
`www.gov.uk` relying party.

To do so, we'll update `govuk_personalisation` with the new link and
change the controller in Frontend that currently serves that page to
redirect on to the new one. This allows us to do an environment aware
redirect.

Before we can update `govuk_personalisation` we need to set these
environment variables. This won't have any effect on where apps send
users at the moment.

[1]: https://github.com/alphagov/govuk_personalisation